safari-600.7-branch/LayoutTests/js/dfg-to-string-on-value-expected.txt

Tests that ToString on a possible-non-cell value works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,42"
PASS "" + foo("foo", i % 2 ? "hello" : 42) is "foo,hello"
PASS successfullyParsed is true

TEST COMPLETE
